requirements.txt
functions-framework==3.*
google-cloud-iam==2.15.2
google-cloud-bigquery==3.13.0
protobuf==4.25.1